requirements.txt
mediapipe==0.10.9
streamlit==1.33.0
streamlit_antd_components==0.3.2
streamlit_modal==0.1.2
google-generativeai==0.4.1
streamlit_mic_recorder==0.0.8
openai==1.16.2
opencv-python==4.9.0.80
pyautogui==0.9.54
numpy==1.26.4